assignment complete

+++ app/controllers/toys_controller.rb
@@ -7,19 +7,29 @@ class ToysController < ApplicationController
   end
 
   def create
-    toy = Toys.create(toy_params)
+    toy = Toys.create!(toy_params)
     render json: toy, status: :created
   end
 
   def update
     toy = Toy.find_by(id: params[:id])
-    toy.update(toy_params)
+    byebug
+    if toy
+      toy.update(toy_params)
+      render json: toy, status: :accepted
+    else
+      render json: { error: "toy not found" }, status: :not_found
+    end
   end
 
   def destroy
     toy = Toy.find_by(id: params[:id])
-    toy.destroy
-    head :no_content
+    if toy
+      toy.destroy
+      head :no_content
+    else 
+      render json: { error: "toy not found" }, status: :not_found
+    end
   end
 
   private
